display files and download.

+++ database/migrations/2018_01_27_133850_create_files_table.php
@@ -19,7 +19,7 @@ class CreateFilesTable extends Migration
             $table->string('file_name');
             $table->integer('type')->unsigned()->nullable();
             $table->integer('class')->unsigned()->nullable();
-            $table->bigInteger('subject')->unsigned()->nullable();
+            $table->integer('subject')->unsigned()->nullable();
             $table->timestamps();
         });
     }
